Properly construct uint64 objects for Coin amounts (#17745)
Properly construct uint64 objects for Coin amounts.

+++ chia/types/coin_spend.py
@@ -13,6 +13,7 @@ from chia.types.blockchain_format.serialized_program import SerializedProgram
 from chia.types.condition_opcodes import ConditionOpcode
 from chia.types.condition_with_args import ConditionWithArgs
 from chia.util.errors import Err, ValidationError
+from chia.util.ints import uint64
 from chia.util.streamable import Streamable, streamable
 
 CoinSpend = chia_rs.CoinSpend
@@ -76,7 +77,7 @@ def compute_additions_with_cost(
         cost += ConditionCost.CREATE_COIN.value
         puzzle_hash = next(atoms).as_atom()
         amount = next(atoms).as_int()
-        ret.append(Coin(parent_id, puzzle_hash, amount))
+        ret.append(Coin(parent_id, puzzle_hash, uint64(amount)))
 
     return ret, cost
